AJ-400 update helper functions and move to separate file
