+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/orch/AsyncImportSpec.scala
@@ -13,11 +13,9 @@ import org.broadinstitute.dsde.workbench.model.ErrorReport
 import org.broadinstitute.dsde.workbench.model.ErrorReportJsonSupport.ErrorReportFormat
 import org.broadinstitute.dsde.workbench.service.{AclEntry, Orchestration, RestException, WorkspaceAccessLevel}
 import org.parboiled.common.FileUtils
-import org.scalatest.Assertion
 import org.scalatest.OptionValues._
 import org.scalatest.concurrent.PatienceConfiguration.{Interval, Timeout}
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
-import org.scalatest.exceptions.TestFailedException
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.time.{Minutes, Seconds, Span}
@@ -25,7 +23,6 @@ import spray.json._
 
 import java.util.UUID
 import java.util.concurrent.TimeUnit
-import javax.xml.parsers.FactoryConfigurationError
 import scala.io.Source
 import scala.util.{Failure, Success, Try}
 
@@ -34,6 +31,7 @@ class AsyncImportSpec
   extends AnyFreeSpec
     with Matchers
     with Eventually
+    with AsyncTestUtils
     with ScalaFutures
     with WorkspaceFixtures
     with Orchestration {
@@ -49,46 +47,6 @@ class AsyncImportSpec
 
   lazy private val expectedEntities: Map[String, EntityTypeMetadata] = Source.fromResource("AsyncImportSpec-pfb-expected-entities.json").getLines().mkString
     .parseJson.convertTo[Map[String, EntityTypeMetadata]]
-
-
-
-  //TODO: These two methods and exception can be put in a utility class for other tests to use
-  class TestEscapedException(
-                              message: String,
-                              failedCodeStackDepth: Int,
-                            ) extends TestFailedException(message, failedCodeStackDepth){
-    def this(message: String) = this(message, 0)
-  }
-
-  /**
-    * Functions as an assertion unless some fail-immediately condition is met,
-    * in which case it throws a test-failing error
-    * @param assertion Scalatest assertion to be checked for success or failure
-    * @param failCondition Condition that, if met, will end an asynchronous test with failure immediately
-    * @param failMessage Message to report in test failure
-    * @return test assertion
-    */
-  def assertWithFailure(assertion: => Assertion, failCondition: Boolean, failMessage: String): Assertion =
-    if (failCondition) throw new FactoryConfigurationError(failMessage)
-    else assertion
-
-  /**
-    * Wraps a function in an eventually wrapped in a try-catch;
-    * Functions as an eventually but will fail test without retrying if test-ending errors thrown in the test function
-    * @param timeout
-    * @param interval
-    * @param testFunction Test to run; should include an assertWithFailure statement
-    * @tparam T the result of the testFunction
-    * @return The result of the testFunction (an assertion) if an error is not thrown
-    */
-  def eventuallyWithFail[T](timeout: Timeout, interval: Interval, testFunction: => T): T =
-    try {
-      eventually(timeout, interval){
-        testFunction
-      }
-    } catch {
-      case ex:FactoryConfigurationError => throw new TestEscapedException(ex.getMessage)
-    }
 
   "Orchestration" - {
 
@@ -110,11 +68,16 @@ class AsyncImportSpec
             }
 
             // poll for completion as owner
-            eventually {
-              val resp: HttpResponse = Orchestration.getRequest( s"${workspaceUrl(projectName, workspaceName)}/importPFB/$importJobId")
-              resp.status shouldBe StatusCodes.OK
-              blockForStringBody(resp).parseJson.asJsObject.fields.get("status").value shouldBe JsString("Done")
-            }
+            eventuallyWithFail{
+            val resp: HttpResponse = Orchestration.getRequest( s"${workspaceUrl(projectName, workspaceName)}/importPFB/$importJobId")
+            resp.status shouldBe StatusCodes.OK
+            val importStatus = blockForStringBody(resp).parseJson.asJsObject.fields.get("status").value
+            assertWithFailure(
+               importStatus shouldBe JsString("Done"),
+              importStatus.equals(JsString("Error")),
+              "Error encountered during import, failing test early"
+            )
+          }
 
             // inspect data entities and confirm correct import as owner
             eventually {
@@ -143,9 +106,14 @@ class AsyncImportSpec
             }
 
             // poll for completion as writer
-            eventually {
+            eventuallyWithFail {
               val resp = Orchestration.getRequest( s"${workspaceUrl(projectName, workspaceName)}/importPFB/$importJobId")(writerToken)
-              blockForStringBody(resp).parseJson.asJsObject.fields.get("status").value shouldBe JsString("Done")
+              val importStatus = blockForStringBody(resp).parseJson.asJsObject.fields.get("status").value
+              assertWithFailure(
+                importStatus shouldBe JsString("Done"),
+                importStatus.equals(JsString("Error")),
+                "Error encountered during import, failing test early"
+              )
             }
 
             // inspect data entities and confirm correct import as writer
